small ui fixes
