--- app/(authenticated)/users/_lib/subscription-constants.ts
@@ -1,5 +1,0 @@
-/**
- * Matches `subscriptions.status` when a user is treated as “Active” on the
- * Users page (Stripe). Keep in sync with billing writes.
- */
-export const USERS_SUBSCRIPTION_STATUS_ACTIVE = "active" as const;
