--- app/(authenticated)/users/_lib/users-table-header.ts
@@ -1,5 +1,0 @@
-/** Sticky table header background (Users page). */
-export const USERS_TABLE_HEADER_BG = "#EDF4FD";
-
-/** Rows per page on the Users table. */
-export const USERS_TABLE_PAGE_SIZE = 10;
